Implement parallel_policy_executor::bulk_then_execute to support launch policies.
