+++ hpx/lcos/detail/sync_implementations.hpp
@@ -7,6 +7,7 @@
 #define HPX_LCOS_SYNC_IMPLEMENTATIONS_JUL_21_2018_0921PM
 
 #include <hpx/config.hpp>
+#include <hpx/lcos/sync_fwd.hpp>
 #include <hpx/lcos/detail/async_implementations.hpp>
 #include <hpx/lcos/detail/sync_implementations_fwd.hpp>
 #include <hpx/runtime/launch_policy.hpp>
